feat: implement interview question generation and analysis, update UI

+++ backend/src/main/java/com/zoop/backend/domain/dto/InterviewScheduleResponseDto.java
@@ -12,8 +12,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class InterviewScheduleResponseDto {
-    private Integer scheduleId;
-    private Integer jobCandidateId;
+    private Long scheduleId;
+    private Long jobCandidateId;
     private LocalDateTime scheduledTime;
     private LocalDateTime deadlineTime;
     private String interviewLink;
